Dismiss the notifications of the changes of the following categories: - clever display
The label "Dismiss the notifications of the changes of the following
categories:" on the page with the change notifications is alway
visible, even if there's nothing to dismiss. It would be helpful that
this label only appears when there's something to dismiss. That is less
confusing, and that makes sense. This fixes
https://github.com/bibledit/bibledit/issues/535.