More logical order of tabs
https://github.com/bibledit/cloud/issues/36
I have used Basic Mode a lot, and the order of the tabs seems
illogical. The main thing I do is go between Translate and Resources.
So it seems they should be next to each other. The third thing that
relates to one verse at a time is Notes. So I would put Notes in the
third place. That leaves two things that are more global in character:
Changes and Settings.