Loads all change notifications despite shaky network connection
This fixes https://github.com/bibledit/cloud/issues/44: Failure to
fetch all change notifications

When the internet connection is sub-standard, the page with the change
notifications fails to fetch, or may fail to fetch, all of the
notifications. Upon a network error, it now stops. Rather, it should
retry on the network error, probably with a small delay. It should
retry to fetch the same as it tried, then continue with the next one.