Clearer user interface
This fixes https://github.com/bibledit/cloud/issues/49: First, the
current menu entry was confusing to me: "Get the installed Resources".
Get them where? Why would I want to 'get' them?
I suggest that the menu entry under the System settings be changed to
Make an archive of all the Resources installed on this device

The last menu entry, for installing Resources could be clearer also:
Import a Bibledit Resources archive into this device.

When the archive has been created, change the screen that appears to
something like:
Download the archive just created

What would help a lot is that after creating the resources archive, the
page would give a summary as to how many resources it found and put
into that archive. And if there's no resource installed locally, it
would give feedback about that also, with the suggestion to install
some, then try again.