More robust consultation notes passage focuser
This fixes https://github.com/bibledit/cloud/issues/179
If a consultation note has two references, for example, it has the current focus Romans 1:1, and another references Romans 2:2. When clicking on that note, the system moves the passage to one of these two references. This may be confusing when the reference is already focused on one of the two. In such a case it would be an improvement that the system notices that it already has one of the two note references in focus, and then does not refocus to the other verse. There have been several cases where this improved behaviour would have led to less confusion.