Fix partial definition
This fixes https://github.com/bibledit/cloud/issues/222.
partial definition of ψυχή

The definition of the Greek dictionary of the word ψυχή is not complete.

Here's what it says, currently:

Strong's 5590 ψυχή psychḗ psoo-khay' from 5594; breath, i.e. (by implication) spirit, abstractly or concretely (the animal sentient principle only; thus distinguished on the one hand from , which is the rational and immortal soul; and on the other from , which is mere vitality, even of plants: these terms thus exactly correspond respectively to the Hebrew , and ); usage in King James Bible:--heart (+ -ily), life, mind, soul, + us, + you.

As can be seen, bits are left out.

Location: Matthew 2:20 (and plenty of others).