Indonesian Free: Disable links to affect personal account
Request: I suggest that the user not be allowed to change the email address or his login. This is a free level, and if they want to change it, all they have to do is start over.
Action point: To disable account manangement so the user cannot change any credentials.
https://github.com/bibledit/cloud/issues/487