Added default styling feature and style improv. Added default styling feature that saves the setting in the current live database_config_user object. It can be configured through the basic and advanced mode. Through the usual personalize/index.html and system/indonesianfree.html.
There are three options:
Basic, which is the basic no-nonsense styling for Bibledit improved for ease of reading.
Light, which is the basic no-nonsense styling with predominantly blue colors.
Dark, which is the same as light but with dark-grey background.

Other changes in this commit is just minor styling improvements, such as:
Interface font changes, sans-serif into Montserrat and serif into Esteban.
Color of the tabs/topbar are now flatout gradient blue.
Color of the book picker are now slightly transparent.
Added hover state of the tabs/topbar.
Gradient green color for the login and signup button.
A few others that's too insignificant to mention.