Latest libmimetic was included
Version 0.9.8 was included because the existing library cannot be compiled with a C++17 compiler due to mimetic using a removed "register" keyword.
So the solution chosen for just now is to include the relevant parts of the mimetic library. Since a newer version of mimetic that supports C++17 is not yet available. https://github.com/bibledit/cloud/issues/719