Remove deprecated libcurl calls
https://github.com/bibledit/cloud/issues/946

The derecated libcurl calls were related to submitting a Bible to the Bible Drop Box.
However it appears that the API of the Bible Drop Box has changed.
Hence it was decided that Bibledit would no longer automatically submit Bibles to the Bible Drop Box.
It is now possible to do that manually only.
As a result of this, all deprecated libcurl calls have gone.