Add ability to reset styles in stylesheet editor
Introduces a reset option in the stylesheet editor UI, allowing users with write access to reset a style to its default values if it is a default style. Updates the changelog and adjusts the view logic and template to support this feature, displaying appropriate options for default and custom styles.
https://github.com/bibledit/cloud/issues/1020